requirements.txt
python-telegram-bot==21.6
requests==2.32.3
sqlalchemy==2.0.36
python-dotenv==1.0.1
apscheduler==3.10.4
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3
